+++ DirectZobEngine/DirectZobWrapper.cpp
@@ -61,7 +61,7 @@ namespace CLI
 	array<System::String^>^ DirectZobWrapper::GetEventsAndClear()
 	{
 		const std::vector<std::string>* data = m_Instance->GetEventManager()->GetEvents();
-		int l = data->size();
+		int l = (int)data->size();
 		array<System::String ^>^ arr = gcnew array<System::String ^>(l);
 		for (int i = 0; i < l; i++)
 		{
